+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Added capability of matching MIDI devices with regex expressions https://github.com/GrandOrgue/grandorgue/issues/885
 - Added a setting option whether to check for existence of active midi devices on startup https://github.com/GrandOrgue/grandorgue/issues/796
 - Added capability of switching auto enabling of new midi devices on/off https://github.com/GrandOrgue/grandorgue/issues/703
 - Added capability of enabling and disabling particular MIDI APIs https://github.com/GrandOrgue/grandorgue/issues/703